In some places I was using u_int_xx and in some places I was using uint_xx so I changed everything to use u_int_xx so it's all consistent.  I also changed the message queue command from int32 to int8, to save a little memory per object, but more so because 0x11111111 is 286331153 in decimal which is less than helpful when you're staring at the value in gdb trying to figure out what kind of message you're looking at.  Using int8 means you'll now see 'P' for packet, 'S' for shutdown.  Spiffy, eh?
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/hades/pkgs@30943 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ src/message.cpp
@@ -140,7 +140,7 @@ aHigh_bytes = high_bytes;
 }
 /*--------------------------------------------------------------------------*/
 /*--------------------------------------------------------------------------*/
-MessageWagon::MessageWagon(int argCommand,const unsigned char *argBuffer,int argLength)
+MessageWagon::MessageWagon(u_int8_t argCommand,const unsigned char *argBuffer,int argLength)
 {
 next = NULL;
 command = argCommand;
@@ -150,7 +150,7 @@ memcpy(buffer,argBuffer,argLength);
 timestamp = time(NULL);
 }
 /*--------------------------------------------------------------------------*/
-MessageWagon::MessageWagon(int argCommand)
+MessageWagon::MessageWagon(u_int8_t argCommand)
 {
 next = NULL;
 command = argCommand;
